weather_app/requirements.txt
requests
geocoder
Pillow